requirements.txt
discord.py==1.7.2
requests==2.24.0
pytz==2020.1
firebase-admin==5.0.1
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.0
google-cloud-firestore==2.1.3
google-cloud-storage==1.39.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
CacheControl==0.12.6
discord-py-slash-command==2.3.1